:hammer: Refactoring code
