--- Pipfile
@@ -1,13 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-pyttsx3 = "*"
-pypdf2 = "*"
-
-[requires]
-python_version = "3.7"
